addressed race condition and moved util function
