+++ src/constants/constants.ts
@@ -119,6 +119,7 @@ export const sqlToolsServiceCrashLink = 'https://github.com/Microsoft/vscode-mss
 export const azureAccountExtensionId = 'ms-vscode.azure-account';
 export const databaseString = 'Database';
 export const localizedTexts = 'localizedTexts';
+export const sqlPlanLanguageId = 'sqlplan';
 export const ipAddressRegex = /\b(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\b/;
 /**
  * Azure Firewall rule name convention is specified here:
