+++ src/reactviews/common/utils.ts
@@ -20,15 +20,3 @@ export function formatString(str: string, ...args: any[]): string {
 	}
 	return result;
 }
-
-/**
- * Generates a random nonce value that can be used in a webview
- */
-export function getNonce(): string {
-	let text = "";
-	const possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-	for (let i = 0; i < 32; i++) {
-		text += possible.charAt(Math.floor(Math.random() * possible.length));
-	}
-	return text;
-}
